LayoutTests/platform/mac/imported/w3c/web-platform-tests/webrtc/RTCRtpReceiver-track-settings.tentative-expected.txt


FAIL RTCRtpReceiver.track.getSettings() contains resolution entries assert_equals: Receive track width setting is zero initially expected (number) 0 but got (undefined) undefined
